back page link

+++ src/components/bookList/BookList.js
@@ -1,24 +1,29 @@
 import React from 'react'
+import { useHistory } from 'react-router-dom'
 import deleteIcon from '../../assets/icons/delete.svg'
 import api from '../../backend/api'
+
 import './styles.css'
 
 function BookList(props){
+  const history = useHistory()
 
   var userId = window.location.href.toString().replace("http://localhost:3000/bookcase/", '')
   var title = props.title
   var author = props.author
 
   async function deleteBook(){
-    console.log(title, ' - ', author)
-    const response = await api.delete(`bookcase/${userId}`, {
+
+    const response = await api.delete(`bookcase/${userId}/${title}/${author}`, {
       params: {
         userId,
         title,
         author
       }
     })
-    console.log(response)
+    if(response.data == true){
+      alert('Livro removido com sucesso.')
+    }
     
   }
   return (
